#278 ログインしていないときにも、likeボタンとbookmarkボタンを押せるようにするため

+++ app/helpers/posts_helper.rb
@@ -1,10 +1,6 @@
 module PostsHelper
   def create_like_button(post_id)
-    unless user_signed_in?
-      return render "shared/icons/heart", class_list: "w-8 h-8 fill-none stroke-current"
-    end
-
-    if current_user.like?(post_id)
+    if user_signed_in? && current_user.like?(post_id)
       button_to post_likes_path(post_id), method: :delete do
         render "shared/icons/heart", class_list: "w-8 h-8 fill-current stroke-current"
       end
@@ -16,11 +12,7 @@ module PostsHelper
   end
 
   def create_bookmark_button(post_id)
-    unless user_signed_in?
-      return render "shared/icons/bookmark", class_list: "w-8 h-8 fill-none stroke-current"
-    end
-
-    if current_user.bookmarking?(post_id)
+    if user_signed_in? && current_user.bookmarking?(post_id)
       button_to post_bookmarks_path(post_id), method: :delete do
         render "shared/icons/bookmark", class_list: "w-8 h-8 fill-current stroke-current"
       end
